Этап 2. Разбиваем проект на микросервисы
Исправление ошибок из тестов рейтингов
Дебагинг UserClient и RatingService
вс feign эндпоинты теперь /system...
Рефакторинг EventService
Все тесты проходят

+++ core/event-service/src/main/java/ru/practicum/ewm/events/controller/system/SystemEventController.java
@@ -21,7 +21,7 @@ public class SystemEventController {
 
 	@GetMapping("/{eventId}")
 	EventFullDto systemFindEventById(@PathVariable @Positive Long eventId,
-	                                 @RequestParam @PositiveOrZero Integer confirmets) {
+	                                 @RequestParam @PositiveOrZero Long confirmets) {
 		return eventService.findEventById(eventId, confirmets);
 	}
 }
